nDisplay: a couple of fixes - fixed broken MU (regression) - changed socket port types from int32 to uint16 - added logs for cluster events
#jira none
#rb Alejandro.Arango

#ROBOMERGE-AUTHOR: andrey.yamashev
#ROBOMERGE-SOURCE: CL 18164713 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v889-18060218)

[CL 18164731 by andrey yamashev in ue5-release-engine-staging branch]

+++ Engine/Plugins/Runtime/nDisplay/Source/DisplayCluster/Private/Cluster/DisplayClusterClusterManager.h
@@ -75,8 +75,8 @@ public:
 	virtual void EmitClusterEventJson(const FDisplayClusterClusterEventJson& Event, bool MasterOnly) override;
 	virtual void EmitClusterEventBinary(const FDisplayClusterClusterEventBinary& Event, bool bMasterOnly) override;
 
-	virtual void SendClusterEventTo(const FString& Address, const int32 Port, const FDisplayClusterClusterEventJson& Event,   bool bMasterOnly) override;
-	virtual void SendClusterEventTo(const FString& Address, const int32 Port, const FDisplayClusterClusterEventBinary& Event, bool bMasterOnly) override;
+	virtual void SendClusterEventTo(const FString& Address, const uint16 Port, const FDisplayClusterClusterEventJson&   Event, bool bMasterOnly) override;
+	virtual void SendClusterEventTo(const FString& Address, const uint16 Port, const FDisplayClusterClusterEventBinary& Event, bool bMasterOnly) override;
 
 public:
 	//////////////////////////////////////////////////////////////////////////////////////////////
